+++ README.md
@@ -14,15 +14,15 @@ Run the project by executing `main.py`:
 python main.py
 ```
 
-This starts the Qt based interface defined in `gui.py` and launches the
-background worker defined in `worker.py` for handling networking and hotkeys.
+This starts the Qt based interface defined in `kvm/gui/gui.py` and launches the
+background worker defined in `kvm/network/worker.py` for handling networking and hotkeys.
 
 ## Building a Windows executable
 
-Use `build_exe.py` to create a standalone executable with PyInstaller:
+Use `tools/build_exe.py` to create a standalone executable with PyInstaller:
 
 ```bash
-python build_exe.py
+python tools/build_exe.py
 ```
 
 The script ensures PyInstaller is installed and bundles `main.py` into a single
@@ -103,17 +103,17 @@ connection to activate immediately.
 
 ## Clipboard synchronization
 
-A simple script `clipboard_sync.py` is provided to share the system clipboard between computers.
+A simple script `tools/clipboard_sync.py` is provided to share the system clipboard between computers.
 One machine runs the server:
 
 ```bash
-python clipboard_sync.py server --port 8765
+python tools/clipboard_sync.py server --port 8765
 ```
 
 Other machines connect as clients:
 
 ```bash
-python clipboard_sync.py client <server-ip> --port 8765
+python tools/clipboard_sync.py client <server-ip> --port 8765
 ```
 
 Whenever the clipboard changes on any connected computer the new contents are sent to all others.
@@ -124,7 +124,7 @@ The script relies on the `pyperclip` library for cross-platform clipboard access
 Code style checks can be performed with `pycodestyle`:
 
 ```bash
-pycodestyle --max-line-length=120 gui.py clipboard_sync.py
+pycodestyle --max-line-length=120 kvm/gui/gui.py tools/clipboard_sync.py
 ```
 
 You can also use the provided configuration in `setup.cfg` to keep
