+++ kvm/gui/gui.py
@@ -32,8 +32,8 @@ from PySide6.QtGui import QIcon, QAction
 from PySide6.QtCore import QSize, QSettings, QThread, Qt, QTimer, QStandardPaths
 
 
-from worker import KVMWorker
-from config import APP_NAME, ORG_NAME, DEFAULT_PORT, ICON_PATH, TEMP_DIR_PARTS
+from ..network.worker import KVMWorker
+from ..config import APP_NAME, ORG_NAME, DEFAULT_PORT, ICON_PATH, TEMP_DIR_PARTS
 
 MB = 1024 * 1024
